support null for get field
